// src/app/shared/model/asset-negative-obj.model.ts
export class AssetNegativeObj {
    AssetNegativeId: any;
    AssetMasterId: any;
    SerialNo1: any;
    SerialNo2: any;
    SerialNo3: any;
    SerialNo4: any;
    SerialNo5: any;
    MrNegAssetSourceCode: any;
    Notes: any;
    IsActive: any;
    RowVersion: any;
    constructor(){this.AssetNegativeId = 0, this.RowVersion = ""}
}